requirements.txt
streamlit==1.32.2

openai>=1.0.0
google-generativeai>=0.5.0

transformers==4.38.2
torch>=2.2.0
sentencepiece>=0.1.99

pillow>=10.0.0
pytesseract>=0.3.10

gtts>=2.5.0
SpeechRecognition>=3.10.0
pydub>=0.25.1

numpy>=1.26.0
requests>=2.31.0